+++ releasenotes.md
@@ -18,6 +18,7 @@
 - Improve error message for missing `$endif`.
 - `foo[x][y] = b` now interpreted as `(*&foo[x])[y] = b` which allows overloads to do chained [] accesses.
 - Error if a stack allocated variable is too big (configurable with `--max-stack-object-size`).
+- Add `@safeinfer` to allow `var` to be used locally.
 
 ### Fixes
 - List.remove_at would incorrectly trigger ASAN.